+++ compiler-core/src/language_server/tests/action.rs
@@ -71,9 +71,11 @@ const CONVERT_TO_USE: &str = "Convert to `use`";
 const EXTRACT_VARIABLE: &str = "Extract variable";
 const EXPAND_FUNCTION_CAPTURE: &str = "Expand function capture";
 const GENERATE_DYNAMIC_DECODER: &str = "Generate dynamic decoder";
+const GENERATE_JSON_ENCODER: &str = "Generate JSON encoder";
 const PATTERN_MATCH_ON_ARGUMENT: &str = "Pattern match on argument";
 const PATTERN_MATCH_ON_VARIABLE: &str = "Pattern match on variable";
 const GENERATE_FUNCTION: &str = "Generate function";
+const CONVERT_TO_FUNCTION_CALL: &str = "Convert to function call";
 
 macro_rules! assert_code_action {
     ($title:expr, $code:literal, $range:expr $(,)?) => {
@@ -4869,5 +4871,258 @@ pub type Names {
 }
 ",
         find_position_of("[").to_selection()
+    );
+}
+
+#[test]
+fn convert_to_function_call_works_with_argument_in_first_position() {
+    assert_code_action!(
+        CONVERT_TO_FUNCTION_CALL,
+        "
+pub fn main() {
+  [1, 2, 3]
+  |> map(todo)
+}
+
+fn map(list: List(a), fun: fn(a) -> b) -> List(b) { todo }
+",
+        find_position_of("map").to_selection()
+    );
+}
+
+#[test]
+fn generate_json_encoder() {
+    assert_code_action!(
+        GENERATE_JSON_ENCODER,
+        "
+pub type Person {
+  Person(name: String, age: Int, height: Float, is_cool: Bool)
+}
+",
+        find_position_of("type").to_selection()
+    );
+}
+
+#[test]
+fn convert_to_function_call_works_with_argument_in_first_position_2() {
+    assert_code_action!(
+        CONVERT_TO_FUNCTION_CALL,
+        "
+pub fn main() {
+  [1, 2, 3] |> wibble
+}
+
+fn wibble(a) { todo }
+",
+        find_position_of("wibble").to_selection()
+    );
+}
+
+#[test]
+fn generate_json_encoder_complex_types() {
+    let src = "
+import gleam/option
+import gleam/dict
+
+pub type Something
+
+pub type Wibble(value) {
+  Wibble(
+    maybe: option.Option(Int),
+    something: Something,
+    map: dict.Dict(String, List(Float)),
+    unknown: List(value),
+  )
+}
+";
+
+    assert_code_action!(
+        GENERATE_JSON_ENCODER,
+        TestProject::for_source(src)
+            .add_module("gleam/option", "pub type Option(a)")
+            .add_module("gleam/dict", "pub type Dict(k, v)"),
+        find_position_of("type W").to_selection()
+    );
+}
+
+#[test]
+fn convert_to_function_call_works_with_argument_in_first_position_3() {
+    assert_code_action!(
+        CONVERT_TO_FUNCTION_CALL,
+        "
+pub fn main() {
+  [1, 2, 3] |> wibble()
+}
+
+fn wibble(a) { todo }
+",
+        find_position_of("wibble").to_selection()
+    );
+}
+
+#[test]
+fn generate_json_encoder_already_imported_module() {
+    let src = "
+import gleam/json as json_encoding
+
+pub type Wibble {
+  Wibble(a: Int, b: Float, c: String)
+}
+";
+
+    assert_code_action!(
+        GENERATE_JSON_ENCODER,
+        TestProject::for_source(src).add_module("gleam/json", "pub type Json"),
+        find_position_of("type W").to_selection()
+    );
+}
+
+#[test]
+fn convert_to_function_call_works_with_argument_in_first_position_4() {
+    assert_code_action!(
+        CONVERT_TO_FUNCTION_CALL,
+        "
+pub fn main() {
+  [1, 2, 3] |> wibble.wobble
+}
+",
+        find_position_of("wibble").to_selection()
+    );
+}
+
+#[test]
+fn generate_json_encoder_tuple() {
+    assert_code_action!(
+        GENERATE_JSON_ENCODER,
+        "
+pub type Wibble {
+  Wibble(tuple: #(Int, Float, #(String, Bool)))
+}
+",
+        find_position_of("type W").to_selection()
+    );
+}
+
+#[test]
+fn convert_to_function_call_works_with_function_producing_another_function() {
+    assert_code_action!(
+        CONVERT_TO_FUNCTION_CALL,
+        "
+pub fn main() {
+  1 |> wibble(2)
+}
+
+fn wibble(c) -> fn(a) -> Nil {
+  fn(_) { Nil }
+}
+",
+        find_position_of("wibble").to_selection()
+    );
+}
+
+#[test]
+fn generate_json_encoder_recursive_type() {
+    let src = "
+import gleam/option.{Some}
+
+pub type LinkedList {
+  LinkedList(value: Int, next: option.Option(LinkedList))
+}
+";
+    assert_code_action!(
+        GENERATE_JSON_ENCODER,
+        TestProject::for_source(src)
+            .add_module("gleam/option", "pub type Option(a) { Some(a) None }"),
+        find_position_of("type").to_selection()
+    );
+}
+
+#[test]
+fn convert_to_function_call_works_with_hole_in_first_position() {
+    assert_code_action!(
+        CONVERT_TO_FUNCTION_CALL,
+        "
+pub fn main() {
+  [1, 2, 3]
+  |> map(_, todo)
+}
+
+fn map(list: List(a), fun: fn(a) -> b) -> List(b) { todo }
+",
+        find_position_of("[").to_selection()
+    );
+}
+
+#[test]
+fn generate_json_encoder_list_of_tuples() {
+    assert_code_action!(
+        GENERATE_JSON_ENCODER,
+        "
+pub type Wibble {
+  Wibble(values: List(#(Int, String)))
+}
+",
+        find_position_of("type").to_selection()
+    );
+}
+
+#[test]
+fn no_code_action_to_generate_json_encoder_for_multi_variant_type() {
+    assert_no_code_actions!(
+        GENERATE_JSON_ENCODER,
+        "
+pub type Wibble {
+  Wibble(wibble: Int)
+  Wobble(wobble: Float)
+}
+",
+        find_position_of("type").to_selection()
+    );
+}
+
+#[test]
+fn convert_to_function_call_works_with_hole_not_in_first_position() {
+    assert_code_action!(
+        CONVERT_TO_FUNCTION_CALL,
+        "
+pub fn main() {
+  fn(a) { todo }
+  |> map([1, 2, 3], _)
+}
+
+fn map(list: List(a), fun: fn(a) -> b) -> List(b) { todo }
+",
+        find_position_of("fn(a)").select_until(find_position_of("map"))
+    );
+}
+
+#[test]
+fn convert_to_function_call_always_inlines_the_first_step() {
+    assert_code_action!(
+        CONVERT_TO_FUNCTION_CALL,
+        "
+pub fn main() {
+  [1, 2, 3]
+  |> map(todo)
+  |> filter(todo)
+}
+
+fn map(list: List(a), fun: fn(a) -> b) -> List(b) { todo }
+fn filter(list: List(a), fun: fn(a) -> Bool) -> List(b) { todo }
+",
+        find_position_of("[1, 2, 3]").select_until(find_position_of("map"))
+    );
+}
+
+#[test]
+fn no_code_action_to_generate_json_encoder_for_type_without_labels() {
+    assert_no_code_actions!(
+        GENERATE_JSON_ENCODER,
+        "
+pub type Wibble {
+  Wibble(Int, Int, String)
+}
+",
+        find_position_of("type").to_selection()
     );
 }