+++ CHANGELOG.md
@@ -13,8 +13,71 @@
   making Gleam packages discoverable through global search of HexDocs.
   ([Diemo Gebhardt](https://github.com/diemogebhardt))
 
+- Allow users to set the `GLEAM_CACERTS_PATH` environment variable to specify a
+  path to a directory containing CA certificates to install Hex packages.
+  ([winstxnhdw](https://github.com/winstxnhdw))
+
 ### Language server
+
+- The language server now offers a code action to convert the first step of a
+  pipeline to a regular function call. For example, this code:
+
+  ```gleam
+  import gleam/list
+
+  pub fn main() {
+    [1, 2, 3] |> list.map(fn(n) { n * 2 })
+  }
+  ```
+
+  Will be rewritten as:
+
+  ```gleam
+  import gleam/list
+
+  pub fn main() {
+    list.map([1, 2, 3], fn(n) { n * 2 })
+  }
+  ```
+
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- The Language Server now suggests a code action to generate a function to
+  encode a custom type as JSON using the `gleam_json` package. For example:
+
+  ```gleam
+  pub type Person {
+    Person(name: String, age: Int)
+  }
+  ```
+
+  Will become:
+
+  ```gleam
+  import gleam/json
+
+  pub type Person {
+    Person(name: String, age: Int)
+  }
+
+  fn encode_person(person: Person) -> json.Json {
+    json.object([
+      #("name", json.string(person.name)),
+      #("age", json.int(person.age)),
+    ])
+  }
+  ```
+
+  ([Surya Rose](https://github.com/GearsDatapacks))
 
 ### Formatter
 
 ### Bug fixes
+
+## v1.8.1 - 2025-02-11
+
+### Bug fixes
+
+- Fixed a metadata caching bug where accessors for opaque types could sometimes
+  be used in other modules.
+  ([Louis Pilfold](https://github.com/lpil))